requirements.txt
Django>=4.2
djangorestframework
requests
python-dateutil
psycopg2